Refactor area control picker into a separate component
