+++ src/panels/lovelace/editor/config-elements/hui-area-controls-card-feature-editor.ts
@@ -1,38 +1,24 @@
-import type { HassEntity } from "home-assistant-js-websocket";
 import { mdiDragHorizontalVariant } from "@mdi/js";
 import { css, html, LitElement, nothing } from "lit";
 import { customElement, property, state } from "lit/decorators";
 import { repeat } from "lit/directives/repeat";
-import Fuse from "fuse.js";
 import memoizeOne from "memoize-one";
 import { fireEvent } from "../../../../common/dom/fire_event";
-import { computeEntityNameList } from "../../../../common/entity/compute_entity_name_display";
 import { computeStateName } from "../../../../common/entity/compute_state_name";
-import { computeRTL } from "../../../../common/util/compute_rtl";
-import type { LocalizeFunc } from "../../../../common/translations/localize";
-import {
-  multiTermSortedSearch,
-  type FuseWeightedKey,
-} from "../../../../resources/fuseMultiTerm";
-import "../../../../components/ha-combo-box-item";
-import "../../../../components/ha-domain-icon";
+import "../../../../components/ha-area-controls-picker";
 import "../../../../components/ha-form/ha-form";
 import type {
   HaFormSchema,
   SchemaUnion,
 } from "../../../../components/ha-form/types";
-import "../../../../components/ha-generic-picker";
-import type { PickerComboBoxItem } from "../../../../components/ha-picker-combo-box";
 import "../../../../components/chips/ha-input-chip";
 import "../../../../components/ha-sortable";
-import "../../../../components/ha-state-icon";
 import "../../../../components/ha-svg-icon";
 import type { HomeAssistant } from "../../../../types";
 import {
-  AREA_CONTROLS_BUTTONS,
   getAreaControlEntities,
   MAX_DEFAULT_AREA_CONTROLS,
-} from "../../card-features/hui-area-controls-card-feature";
+} from "../../../../data/area_controls";
 import {
   AREA_CONTROL_DOMAINS,
   type AreaControl,
@@ -41,13 +27,6 @@ import {
 } from "../../card-features/types";
 import type { AreaCardFeatureContext } from "../../cards/hui-area-card";
 import type { LovelaceCardFeatureEditor } from "../../types";
-
-interface AreaControlPickerItem extends PickerComboBoxItem {
-  type?: "domain" | "entity";
-  stateObj?: HassEntity;
-  domain?: string;
-  deviceClass?: string;
-}
 
 type AreaControlsCardFeatureData = AreaControlsCardFeatureConfig & {
   customize_controls: boolean;
@@ -99,191 +78,6 @@ export class HuiAreaControlsCardFeatureEditor
         Object.keys(controlEntities) as (keyof typeof controlEntities)[]
       ).filter((control) => controlEntities[control].length > 0);
     }
-  );
-
-  private _domainSearchKeys: FuseWeightedKey[] = [
-    {
-      name: "primary",
-      weight: 10,
-    },
-  ];
-
-  private _entitySearchKeys: FuseWeightedKey[] = [
-    {
-      name: "primary",
-      weight: 10,
-    },
-    {
-      name: "secondary",
-      weight: 5,
-    },
-    {
-      name: "id",
-      weight: 3,
-    },
-  ];
-
-  private _createFuseIndex = (
-    items: AreaControlPickerItem[],
-    keys: FuseWeightedKey[]
-  ) => Fuse.createIndex(keys, items);
-
-  private _domainFuseIndex = memoizeOne((items: AreaControlPickerItem[]) =>
-    this._createFuseIndex(items, this._domainSearchKeys)
-  );
-
-  private _entityFuseIndex = memoizeOne((items: AreaControlPickerItem[]) =>
-    this._createFuseIndex(items, this._entitySearchKeys)
-  );
-
-  private _getItems = memoizeOne(
-    (
-      areaId: string,
-      excludeEntities: string[] | undefined,
-      currentValue: AreaControl[],
-      localize: LocalizeFunc,
-      _entities: HomeAssistant["entities"],
-      _devices: HomeAssistant["devices"],
-      _areas: HomeAssistant["areas"]
-    ): ((
-      searchString?: string,
-      section?: string
-    ) => (AreaControlPickerItem | string)[]) =>
-      (searchString?: string, section?: string) => {
-        if (!this.hass) {
-          return [];
-        }
-
-        const isSelected = (id: string): boolean =>
-          currentValue.some((item) =>
-            typeof item === "string" ? item === id : item.entity_id === id
-          );
-
-        const controlEntities = getAreaControlEntities(
-          AREA_CONTROL_DOMAINS as unknown as AreaControlDomain[],
-          areaId,
-          excludeEntities,
-          this.hass
-        );
-
-        const items: (AreaControlPickerItem | string)[] = [];
-        let domainItems: AreaControlPickerItem[] = [];
-        let entityItems: AreaControlPickerItem[] = [];
-
-        if (!section || section === "domain") {
-          const supportedControls = (
-            Object.keys(controlEntities) as (keyof typeof controlEntities)[]
-          ).filter((control) => controlEntities[control].length > 0);
-
-          supportedControls.forEach((control) => {
-            if (isSelected(control)) {
-              return;
-            }
-            const label = localize(
-              `ui.panel.lovelace.editor.features.types.area-controls.controls_options.${control}`
-            );
-            const button = AREA_CONTROLS_BUTTONS[control];
-            const deviceClass = button.filter.device_class
-              ? Array.isArray(button.filter.device_class)
-                ? button.filter.device_class[0]
-                : button.filter.device_class
-              : undefined;
-
-            domainItems.push({
-              type: "domain",
-              id: control,
-              primary: label,
-              domain: button.filter.domain,
-              deviceClass,
-            });
-          });
-
-          if (searchString) {
-            const fuseIndex = this._domainFuseIndex(domainItems);
-            domainItems = multiTermSortedSearch(
-              domainItems,
-              searchString,
-              this._domainSearchKeys,
-              (item) => item.id,
-              fuseIndex
-            );
-          }
-        }
-
-        if (!section || section === "entity") {
-          const allEntityIds = Object.values(controlEntities).flat();
-          const uniqueEntityIds = Array.from(new Set(allEntityIds));
-
-          const isRTL = computeRTL(this.hass);
-
-          uniqueEntityIds.forEach((entityId) => {
-            if (isSelected(entityId)) {
-              return;
-            }
-            const stateObj = this.hass!.states[entityId];
-            if (!stateObj) {
-              return;
-            }
-
-            const [entityName, deviceName, areaName] = computeEntityNameList(
-              stateObj,
-              [{ type: "entity" }, { type: "device" }, { type: "area" }],
-              this.hass!.entities,
-              this.hass!.devices,
-              this.hass!.areas,
-              this.hass!.floors
-            );
-
-            const primary = entityName || deviceName || entityId;
-            const secondary = [areaName, entityName ? deviceName : undefined]
-              .filter(Boolean)
-              .join(isRTL ? " ◂ " : " ▸ ");
-
-            entityItems.push({
-              type: "entity",
-              id: entityId,
-              primary,
-              secondary,
-              stateObj,
-            });
-          });
-
-          if (searchString) {
-            const fuseIndex = this._entityFuseIndex(entityItems);
-            entityItems = multiTermSortedSearch(
-              entityItems,
-              searchString,
-              this._entitySearchKeys,
-              (item) => item.id,
-              fuseIndex
-            );
-          }
-        }
-
-        // Only add section headers if there are items in that section
-        if (!section) {
-          if (domainItems.length > 0) {
-            items.push(
-              localize(
-                "ui.panel.lovelace.editor.features.types.area-controls.sections.domain"
-              )
-            );
-            items.push(...domainItems);
-          }
-          if (entityItems.length > 0) {
-            items.push(
-              localize(
-                "ui.panel.lovelace.editor.features.types.area-controls.sections.entity"
-              )
-            );
-            items.push(...entityItems);
-          }
-        } else {
-          items.push(...domainItems, ...entityItems);
-        }
-
-        return items;
-      }
   );
 
   protected render() {
@@ -360,70 +154,20 @@ export class HuiAreaControlsCardFeatureEditor
                   </ha-sortable>
                 `
               : nothing}
-            <ha-generic-picker
+            <ha-area-controls-picker
               .hass=${this.hass}
+              .areaId=${this.context.area_id}
+              .excludeEntities=${this.context.exclude_entities}
               .value=${""}
               .addButtonLabel=${this.hass.localize(
                 "ui.panel.lovelace.editor.features.types.area-controls.controls"
               )}
-              .getItems=${this._getItems(
-                this.context.area_id,
-                this.context.exclude_entities,
-                value,
-                this.hass.localize,
-                this.hass.entities,
-                this.hass.devices,
-                this.hass.areas
-              )}
-              .rowRenderer=${this._rowRenderer as any}
-              .sections=${[
-                {
-                  id: "domain",
-                  label: this.hass.localize(
-                    "ui.panel.lovelace.editor.features.types.area-controls.sections.domain"
-                  ),
-                },
-                {
-                  id: "entity",
-                  label: this.hass.localize(
-                    "ui.panel.lovelace.editor.features.types.area-controls.sections.entity"
-                  ),
-                },
-              ]}
               @value-changed=${this._controlChanged}
-            ></ha-generic-picker>
+            ></ha-area-controls-picker>
           `
         : nothing}
     `;
   }
-
-  private _rowRenderer = (item: AreaControlPickerItem) => html`
-    <ha-combo-box-item type="button" compact>
-      ${item.type === "entity" && item.stateObj
-        ? html`<ha-state-icon
-            slot="start"
-            .hass=${this.hass}
-            .stateObj=${item.stateObj}
-          ></ha-state-icon>`
-        : item.domain
-          ? html`<ha-domain-icon
-              slot="start"
-              .hass=${this.hass}
-              .domain=${item.domain}
-              .deviceClass=${item.deviceClass}
-            ></ha-domain-icon>`
-          : nothing}
-      <span slot="headline">${item.primary}</span>
-      ${item.secondary
-        ? html`<span slot="supporting-text">${item.secondary}</span>`
-        : nothing}
-      ${item.type === "entity" && item.stateObj
-        ? html`<span slot="supporting-text" class="code">
-            ${item.stateObj.entity_id}
-          </span>`
-        : nothing}
-    </ha-combo-box-item>
-  `;
 
   private _getItemLabel(item: AreaControl): string {
     if (!this.hass) {
@@ -528,9 +272,6 @@ export class HuiAreaControlsCardFeatureEditor
     }
     ha-chip-set {
       margin-bottom: var(--ha-space-2);
-    }
-    .code {
-      font-family: var(--ha-font-family-code);
     }
   `;
 }
