+++ src/panels/lovelace/strategies/areas/areas-overview-view-strategy.ts
@@ -1,10 +1,10 @@
 import { ReactiveElement } from "lit";
 import { customElement } from "lit/decorators";
 import { floorDefaultIcon } from "../../../../components/ha-floor-icon";
+import { getAreaControlEntities } from "../../../../data/area_controls";
 import type { LovelaceSectionConfig } from "../../../../data/lovelace/config/section";
 import type { LovelaceViewConfig } from "../../../../data/lovelace/config/view";
 import type { HomeAssistant } from "../../../../types";
-import { getAreaControlEntities } from "../../card-features/hui-area-controls-card-feature";
 import {
   AREA_CONTROL_DOMAINS,
   type AreaControlDomain,
